# backend/routes/api.py
"""
backend/routes/api.py
=====================
API routes for Recall.
Provides endpoints for items, search, graph visualization, quizzes, reminders, and Drive sync.
All endpoints require bearerAuth or telegramInitData (applied via OpenAPI customizer).
"""

import logging
from typing import List, Optional
from fastapi import APIRouter, Path, Query, Response, status
from pydantic import BaseModel, Field

from backend.models.schemas import (
    ItemResponse,
    ItemCreateRequest,
    SearchRequest,
    SearchResponse,
    GraphResponse,
    QuizResponse,
    QuizAnswerRequest,
    QuizStatsResponse,
    ReminderResponse,
    ReminderCreateRequest,
    ErrorResponse,
)

logger = logging.getLogger(__name__)
router = APIRouter(prefix="/api", tags=["api"])

# ---------------------------------------------------------------------------
# Items Group
# ---------------------------------------------------------------------------
@router.get(
    "/items",
    response_model=List[ItemResponse],
    tags=["items"],
    summary="Get saved items",
    description="Returns a paginated list of saved items for the authenticated user, optionally filtered by tag.",
    responses={401: {"model": ErrorResponse}},
)
async def get_items(
    tag: Optional[str] = Query(None, description="Optional tag to filter items by."),
    page: int = Query(1, ge=1, description="Page number for pagination."),
):
    """Retrieve saved items with pagination."""
    return []

@router.post(
    "/items",
    response_model=ItemResponse,
    status_code=status.HTTP_201_CREATED,
    tags=["items"],
    summary="Save a new item",
    description="Saves a new item (url with optional title) for the authenticated user.",
    responses={401: {"model": ErrorResponse}},
)
async def create_item(req: ItemCreateRequest):
    """Save a new item."""
    from datetime import datetime, timezone
    return {
        "id": 1,
        "user_id": 1,
        "source_type": "url",
        "source_url": req.url,
        "summary": "Stub summary",
        "title": req.title or "Stub title",
        "tags": [],
        "created_at": datetime.now(timezone.utc),
    }

@router.delete(
    "/items/{item_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    tags=["items"],
    summary="Delete an item",
    description="Deletes a saved item by ID. Validates ownership to prevent IDOR.",
    responses={
        401: {"model": ErrorResponse},
        404: {"model": ErrorResponse, "description": "Item not found."},
    },
)
async def delete_item(
    item_id: int = Path(..., description="ID of the item to delete."),
):
    """Delete an item."""
    return Response(status_code=status.HTTP_204_NO_CONTENT)

# ---------------------------------------------------------------------------
# Search Group
# ---------------------------------------------------------------------------
@router.post(
    "/search",
    response_model=SearchResponse,
    tags=["search"],
    summary="Search items",
    description="Performs a hybrid (vector + text trigram GIN) search on the user's saved items.",
    responses={401: {"model": ErrorResponse}},
)
async def search_items(req: SearchRequest):
    """Search items."""
    return {"results": []}

# ---------------------------------------------------------------------------
# Graph Group
# ---------------------------------------------------------------------------
@router.get(
    "/graph",
    response_model=GraphResponse,
    tags=["graph"],
    summary="Get mind map graph",
    description="Returns nodes (items, semantic hubs) and edges (similarity links) for graph visualization.",
    responses={401: {"model": ErrorResponse}},
)
async def get_graph():
    """Retrieve mind map graph."""
    return {"nodes": [], "edges": []}

# ---------------------------------------------------------------------------
# Quizzes Group
# ---------------------------------------------------------------------------
@router.get(
    "/quizzes/due",
    response_model=List[QuizResponse],
    tags=["quizzes"],
    summary="Get due quizzes",
    description="Returns a list of quizzes due for review based on SM-2 scheduling.",
    responses={401: {"model": ErrorResponse}},
)
async def get_due_quizzes():
    """Get due quizzes."""
    return []

@router.post(
    "/quizzes/{id}/answer",
    response_model=QuizResponse,
    tags=["quizzes"],
    summary="Submit quiz answer",
    description="Records response quality (0-5) and updates SM-2 scheduling parameters for the quiz.",
    responses={
        401: {"model": ErrorResponse},
        404: {"model": ErrorResponse, "description": "Quiz not found."},
    },
)
async def answer_quiz(
    id: int = Path(..., description="Quiz ID."),
    req: QuizAnswerRequest = ...,
):
    """Answer quiz."""
    from datetime import datetime, timezone, date
    return {
        "id": id,
        "user_id": 1,
        "item_id": 1,
        "question": "Stub question?",
        "options": ["A", "B", "C", "D"],
        "correct_index": 0,
        "explanation": "Stub explanation",
        "ease_factor": 2.5,
        "interval_days": 1,
        "next_review": date.today(),
        "created_at": datetime.now(timezone.utc),
    }

@router.get(
    "/quizzes/stats",
    response_model=QuizStatsResponse,
    tags=["quizzes"],
    summary="Get quiz stats",
    description="Returns aggregated quiz statistics (total, due, reviews, average ease, streak) for the user.",
    responses={401: {"model": ErrorResponse}},
)
async def get_quiz_stats():
    """Get quiz statistics."""
    return {
        "total_quizzes": 0,
        "due_today": 0,
        "completed_reviews": 0,
        "average_ease_factor": 2.5,
        "streak": 0,
    }

# ---------------------------------------------------------------------------
# Reminders Group
# ---------------------------------------------------------------------------
@router.get(
    "/reminders",
    response_model=List[ReminderResponse],
    tags=["reminders"],
    summary="Get reminders",
    description="Returns all reminders configured by the user (up to 20 limit).",
    responses={401: {"model": ErrorResponse}},
)
async def get_reminders():
    """Get reminders."""
    return []

@router.post(
    "/reminders",
    response_model=ReminderResponse,
    status_code=status.HTTP_201_CREATED,
    tags=["reminders"],
    summary="Create reminder",
    description="Saves a new reminder message scheduled for a specific UTC timestamp.",
    responses={401: {"model": ErrorResponse}},
)
async def create_reminder(req: ReminderCreateRequest):
    """Create a new reminder."""
    from datetime import datetime, timezone
    return {
        "id": 1,
        "user_id": 1,
        "item_id": req.item_id,
        "message": req.message,
        "remind_at": req.remind_at,
        "status": "pending",
        "created_at": datetime.now(timezone.utc),
    }

@router.delete(
    "/reminders/{id}",
    status_code=status.HTTP_204_NO_CONTENT,
    tags=["reminders"],
    summary="Delete a reminder",
    description="Deletes a reminder by ID. Validates ownership to prevent IDOR.",
    responses={
        401: {"model": ErrorResponse},
        404: {"model": ErrorResponse, "description": "Reminder not found."},
    },
)
async def delete_reminder(
    id: int = Path(..., description="Reminder ID to delete."),
):
    """Delete a reminder."""
    return Response(status_code=status.HTTP_204_NO_CONTENT)

# ---------------------------------------------------------------------------
# Drive Group
# ---------------------------------------------------------------------------
@router.post(
    "/drive/sync",
    status_code=status.HTTP_202_ACCEPTED,
    tags=["drive"],
    summary="Sync Google Drive",
    description="Triggers a background synchronization of Recall items to Google Drive as a unified doc.",
    responses={
        401: {"model": ErrorResponse},
        429: {"model": ErrorResponse, "description": "Sync limit exceeded (max 5 requests per hour)."},
    },
)
async def sync_drive():
    """Sync items to Google Drive."""
    return {"status": "ok"}

@router.delete(
    "/drive",
    status_code=status.HTTP_204_NO_CONTENT,
    tags=["drive"],
    summary="Disconnect Google Drive",
    description="Clears the user's stored Google refresh token, disconnecting Google Drive integration.",
    responses={401: {"model": ErrorResponse}},
)
async def disconnect_drive():
    """Disconnect Google Drive integration."""
    return Response(status_code=status.HTTP_204_NO_CONTENT)